Add support for mysql
We add an implementation for SpecificQueryProvider for mysql.
We capitalize all references to tables in functions that are shared
between mysql and postgresql.
We replace ILIKE by LIKE throughout for shared functionality, as the
postgresql implementation for SpecificQueryProvider already contains
operations such as "lower(field) [operator] lower('value')".

The mysql queries are targeted at mysql < 8.0.0, i.e. the `WITH query
AS name; SELECT COUNT(*) FROM name` statements are replaced by `SELECT
COUNT(*) FROM (query)`. The other distinctions are the concat operator
`||` which is replaced by the CONCAT function, and casting to decimal
instead of bigint.

Signed-off-by: Peter Verraedt <peter.verraedt@kuleuven.be>