Remove LOWER functions on file properties/metadata
Database queries for metadata or file properties cannot use indices if
the function LOWER is called on the select query. We therefore remove
the LOWER functions altogether. This is a breaking change, making
metadata queries case sensitive, but this copies the behaviour of iquest
calls or wildcard behaviour of imeta.

Signed-off-by: Peter Verraedt <peter.verraedt@kuleuven.be>